feat: 레이싱에 참가한 차량들은 이름이 중복될 수 없다
