+++ kotlin-playground/README.md
@@ -23,4 +23,8 @@ UI
 - 인풋 
 - 예외 
   - 이름은 쉼표로 구분하여 판단 
+
+차량 이름 
+- 이름은 중복될 수 없다 
+- 이름은 5글자 초과이다
