+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/domain/car/exception/DuplicatedCarNameException.kt
@@ -1,0 +1,3 @@
+package org.gonza.kotlinplayground.domain.car.exception
+
+class DuplicatedCarNameException : RuntimeException("차량 이름이 중복되었습니다")
